pkg/pulumi/model/types_test.go
package model

import (
	"reflect"
	"sync"
	"testing"
)

func TestNewResourceTracker(t *testing.T) {
	tests := []struct {
		name string
		want *ResourceTracker
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := NewResourceTracker(); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("NewResourceTracker() = %v, want %v", got, tt.want)
			}
		})
	}
}

func TestResourceTracker_AddResource(t *testing.T) {
	type fields struct {
		resources map[string]map[string]interface{}
		metadata  map[string]map[string]ResourceMetadata
		mu        sync.RWMutex
	}
	type args struct {
		resourceType string
		name         string
		resource     interface{}
		dependencies []string
	}
	tests := []struct {
		name    string
		fields  fields
		args    args
		wantErr bool
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			rt := &ResourceTracker{
				resources: tt.fields.resources,
				metadata:  tt.fields.metadata,
				mu:        tt.fields.mu,
			}
			if err := rt.AddResource(tt.args.resourceType, tt.args.name, tt.args.resource, tt.args.dependencies...); (err != nil) != tt.wantErr {
				t.Errorf("ResourceTracker.AddResource() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestResourceTracker_UpdateResourceState(t *testing.T) {
	type fields struct {
		resources map[string]map[string]interface{}
		metadata  map[string]map[string]ResourceMetadata
		mu        sync.RWMutex
	}
	type args struct {
		resourceType string
		name         string
		state        ResourceState
	}
	tests := []struct {
		name    string
		fields  fields
		args    args
		wantErr bool
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			rt := &ResourceTracker{
				resources: tt.fields.resources,
				metadata:  tt.fields.metadata,
				mu:        tt.fields.mu,
			}
			if err := rt.UpdateResourceState(tt.args.resourceType, tt.args.name, tt.args.state); (err != nil) != tt.wantErr {
				t.Errorf("ResourceTracker.UpdateResourceState() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestResourceTracker_GetResourceState(t *testing.T) {
	type fields struct {
		resources map[string]map[string]interface{}
		metadata  map[string]map[string]ResourceMetadata
		mu        sync.RWMutex
	}
	type args struct {
		resourceType string
		name         string
	}
	tests := []struct {
		name    string
		fields  fields
		args    args
		want    ResourceState
		wantErr bool
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			rt := &ResourceTracker{
				resources: tt.fields.resources,
				metadata:  tt.fields.metadata,
				mu:        tt.fields.mu,
			}
			got, err := rt.GetResourceState(tt.args.resourceType, tt.args.name)
			if (err != nil) != tt.wantErr {
				t.Errorf("ResourceTracker.GetResourceState() error = %v, wantErr %v", err, tt.wantErr)
				return
			}
			if got != tt.want {
				t.Errorf("ResourceTracker.GetResourceState() = %v, want %v", got, tt.want)
			}
		})
	}
}

func TestResourceTracker_GetDependencies(t *testing.T) {
	type fields struct {
		resources map[string]map[string]interface{}
		metadata  map[string]map[string]ResourceMetadata
		mu        sync.RWMutex
	}
	type args struct {
		resourceType string
		name         string
	}
	tests := []struct {
		name    string
		fields  fields
		args    args
		want    ResourceDependency
		wantErr bool
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			rt := &ResourceTracker{
				resources: tt.fields.resources,
				metadata:  tt.fields.metadata,
				mu:        tt.fields.mu,
			}
			got, err := rt.GetDependencies(tt.args.resourceType, tt.args.name)
			if (err != nil) != tt.wantErr {
				t.Errorf("ResourceTracker.GetDependencies() error = %v, wantErr %v", err, tt.wantErr)
				return
			}
			if !reflect.DeepEqual(got, tt.want) {
				t.Errorf("ResourceTracker.GetDependencies() = %v, want %v", got, tt.want)
			}
		})
	}
}

func TestResourceTracker_GetResource(t *testing.T) {
	type fields struct {
		resources map[string]map[string]interface{}
		metadata  map[string]map[string]ResourceMetadata
		mu        sync.RWMutex
	}
	type args struct {
		resourceType string
		name         string
	}
	tests := []struct {
		name   string
		fields fields
		args   args
		want   interface{}
		want1  bool
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			rt := &ResourceTracker{
				resources: tt.fields.resources,
				metadata:  tt.fields.metadata,
				mu:        tt.fields.mu,
			}
			got, got1 := rt.GetResource(tt.args.resourceType, tt.args.name)
			if !reflect.DeepEqual(got, tt.want) {
				t.Errorf("ResourceTracker.GetResource() got = %v, want %v", got, tt.want)
			}
			if got1 != tt.want1 {
				t.Errorf("ResourceTracker.GetResource() got1 = %v, want %v", got1, tt.want1)
			}
		})
	}
}

func TestResourceTracker_GetResourcesByType(t *testing.T) {
	type fields struct {
		resources map[string]map[string]interface{}
		metadata  map[string]map[string]ResourceMetadata
		mu        sync.RWMutex
	}
	type args struct {
		resourceType string
	}
	tests := []struct {
		name   string
		fields fields
		args   args
		want   map[string]interface{}
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			rt := &ResourceTracker{
				resources: tt.fields.resources,
				metadata:  tt.fields.metadata,
				mu:        tt.fields.mu,
			}
			if got := rt.GetResourcesByType(tt.args.resourceType); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("ResourceTracker.GetResourcesByType() = %v, want %v", got, tt.want)
			}
		})
	}
}

func TestResourceTracker_DeleteResource(t *testing.T) {
	type fields struct {
		resources map[string]map[string]interface{}
		metadata  map[string]map[string]ResourceMetadata
		mu        sync.RWMutex
	}
	type args struct {
		resourceType string
		name         string
	}
	tests := []struct {
		name    string
		fields  fields
		args    args
		wantErr bool
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			rt := &ResourceTracker{
				resources: tt.fields.resources,
				metadata:  tt.fields.metadata,
				mu:        tt.fields.mu,
			}
			if err := rt.DeleteResource(tt.args.resourceType, tt.args.name); (err != nil) != tt.wantErr {
				t.Errorf("ResourceTracker.DeleteResource() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestResourceTracker_ListResourceTypes(t *testing.T) {
	type fields struct {
		resources map[string]map[string]interface{}
		metadata  map[string]map[string]ResourceMetadata
		mu        sync.RWMutex
	}
	tests := []struct {
		name   string
		fields fields
		want   []string
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			rt := &ResourceTracker{
				resources: tt.fields.resources,
				metadata:  tt.fields.metadata,
				mu:        tt.fields.mu,
			}
			if got := rt.ListResourceTypes(); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("ResourceTracker.ListResourceTypes() = %v, want %v", got, tt.want)
			}
		})
	}
}
